refactor: Syntax nodes have a script, Tokens don't
The goal is to ensure a new data structure invariant.
Every node in the Syntax tree has a corresponding text range
that begins and ends in the same Script.
When a syntax error is reported, the IDE can highlight a
single text range in a single script, to indicate where the
error is located.

Also, Token and Syntax objects are now smaller.
Tokens are 50% smaller, for a slight speedup in lexical analysis.