simplified lexical grammar
The lexical grammar for numerals has been simplified so that you
can use a 'greedy' tokenizer to segment Curv programs into tokens:
scan left to right, and each point, take the longest substring that
matches a token.

The actual change is: numerals like '1.' with a trailing decimal point
are no longer allowed. This eliminates a conflict with the '..' token,
in a phrase like '1..10'. In practice, I would never write '1.', I would
always just write '1' instead. So I feel that nothing important is lost.

This simplification made it easier to implement syntax colouring
for the 'gedit' text editor.