fix dangling pointer bug
The bug manifested in the REPL after I replaced readline with replxx.
Readline used to return a fresh copy of the command line just typed,
while replxx returns a reference to an internal buffer that gets overwritten.
Class C_Source_String had an unsafe interface, so I got rid of it.