`where`: grammar change: a where phrase is a list item
The precedence of the `where` operator is now higher than `,` or `;`
but lower than the low-precedence right associative operators.
Split the non-terminal `item` into two: `item` now just contains `where`,
and `ritem` contains what used to be called `item`.

Previously, it has been necessary to parenthesize all uses of `where` in
the standard library, but that is going to change. Now, the grammar lets
you append a `where` clause to any `ritem` without adding non-local parentheses.