'parametric' adds a #call field to parametric records
The call field contains the constructor.
This overrides any previous call field (eg, if the record being converted
was already parametric on different parameters). The call field is used
to rerun the constructor with different parameter values.
(The previous scheme was too complex, had inconsistent behaviour,
and was not something that I wanted to document.)